examples/strands-streaming/pyproject.toml
[project]
name = "strands-streaming-agent"
version = "0.1.0"
description = "Strands AI agent with async streaming for AWS Bedrock AgentCore"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=0.1.0",
    "bedrock-agentcore>=0.1.7",
]
